feat: StepHeader prev, next click 함수 추가 (#31)
