+++ src/pages/myPage/ui/customize/step1/components/StepHeader.tsx
@@ -1,3 +1,5 @@
+import { useNavigate } from 'react-router-dom'
+
 import styled from 'styled-components'
 
 import { Button, Header, SvgButton } from '@shared/ui'
@@ -8,20 +10,36 @@ import { flexRowCenter } from '@/shared/styles/mixins'
 
 const StepHeader = ({
   currentStep,
-  onPrevClick,
-  onNextClick,
+  setCurrentStep,
   isValidate,
 }: {
   currentStep: CustomizeStep
-  onPrevClick: () => void
-  onNextClick: () => void
+  setCurrentStep: (step: CustomizeStep) => void
   isValidate: boolean
 }) => {
+  const navigate = useNavigate()
+
+  // current step별 header 뒤로가기 로직
+  const onHeaderPrevClick = () => {
+    if (currentStep === 1 || !currentStep) {
+      navigate('/mypage')
+      return
+    }
+    setCurrentStep((currentStep - 1) as CustomizeStep)
+  }
+
+  // current step별 header 다음 버튼 로직
+  const onHeaderNextClick = () => {
+    if (isValidate) {
+      setCurrentStep((currentStep - 1) as CustomizeStep)
+    }
+  }
+
   return (
     <Header
       left={
         <HeaderLeftWrap>
-          <SvgButton icon={LeftArrow} width={24} height={24} onClick={onPrevClick} />
+          <SvgButton icon={LeftArrow} width={24} height={24} onClick={onHeaderPrevClick} />
           <StepContainer>
             <StepItem $isActive={currentStep === 1}>1</StepItem>
             <StepItem $isActive={currentStep === 2}>2</StepItem>
@@ -30,7 +48,7 @@ const StepHeader = ({
       }
       right={
         <HeaderRight>
-          <Button state={isValidate ? 'primary' : 'disabled'} size="S" onClick={onNextClick}>
+          <Button state={isValidate ? 'primary' : 'disabled'} size="S" onClick={onHeaderNextClick}>
             {currentStep === 1 ? '다음' : '저장'}
           </Button>
         </HeaderRight>
